Change test name to markup
